Bugzilla#685 - rsync service is not work
[Bug/Feature]
Bug

[Root Causes]
in sonic 202211, it will install rsync on all dockers.
And it will install rsync on host due to it will be called in
sonic-build-hooks. After installation, host will have one more service-
rsync.service.

The rsync.service need /etc/rsync.conf as critical file. we don't have
it. so it always be inactived.

[Solution]
There have a PR for master which mentioned we don't need to install
rsync if vcached is disabled.
https://github.com/sonic-net/sonic-buildimage/pull/14407

Since we are not using rsync in host. Turn off as workaround.